test: add static-analysis gate (clang-tidy/cppcheck/qmllint) + fix 8 analyzer findings
Adds the project's missing static-analysis layer: .clang-tidy (curated defect-focused
set), .cppcheck-suppressions, scripts/lint.ps1 + scripts/lint.sh, and a lint CI
workflow; CMAKE_EXPORT_COMPILE_COMMANDS forced ON. clang-tidy gate is clean (143 TUs,
0 findings); qmllint is informational (can't resolve C++ QML types under no-qml-module).

Fixes it surfaced: expLevelRangePercent integer-division-as-float (+ div0 guard,
regression test); unguarded toType1->ind deref in PokemonBox::update; PokemonMove ctor
dead store; signed/unsigned char compare; int-multiply widening; const-ref loop fixes;
PokedexModel switch default. ~ItemMarketEntry pure-virtual-call suppressed + flagged.

Full ctest 67/67 green. VERSION 0.14.0-alpha -> 0.14.1-alpha.

+++ projects/savefile/src/pse-savefile/expanded/world/worldscripts.cpp
@@ -97,7 +97,7 @@ void WorldScripts::scriptsSet(int ind, int val)
 
 void WorldScripts::reset()
 {
-  memset(curScripts, 0, scriptCount * 2); // 2 * 8-bit widths
+  memset(curScripts, 0, static_cast<size_t>(scriptCount) * 2); // 2 * 8-bit widths
   curScriptsChanged();
 }
